Increase default stack size
pthreads has a default stack of 8MB, so it seems reasonable to use the same default for a fiber.

Reduced the number of guard pages to a single page. The PHP VM rarely allocates a large amount of memory at once, and tends to access it in order, so a jumping over a guard region seems unlikely.