Fix Windows build
This is a super hacky solution, but should work for any shared builds.
First we need to fix the obj filenames, since these are expected in the
src/ subfolder.  To trigger their compilation, we write the required
dependency rules manually to the Makefile.  Finally we add the assembler
obj files as LD_FLAGS.

A clean solution would be to support asm files as source files in the
Windows build chain.  If that does not happen, at least we should use
a Makefile.fragment, to avoid the dependency on `MFO`.  We also should
strive to minimize the use of global variables in config.w32.